Avoid specifying parent activities in the AndroidManifest
This is potentially risky as most Activities require some arguments which are not supplied if Android starts the Activity automatically.
Another benefit of this is, that when the user clicks on a chat notification and navigates back, we can now show the ConferencesFragment instead of the NewsFragment, automatically chosen by Android.