Make manga chapters reside in their own directory
Manga pages (offline or cached) are now downloaded into the internalFilesDir/manga/entryId/chapterId directory. This avoids them to be cleared by Glide due to the 250MB limit and retains them also on automatic cache clear, performed by either android due to low memory or manually by the user.
This furthermore allows to actually delete the images when the user deletes a chapter.
Another improvement is the lower memory consumption in the reader, since the images are not hold in the Glide cache anymore. This was a waste of memory, as they were loaded from disk nevertheless.